Food_Ordering_App/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extensions==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
pydot==1.4.2
pygeoip==0.3.2
PyJWT==2.0.1
pyotp==2.6.0
pyparsing==2.4.7
pytz==2021.1
sqlparse==0.4.1
